+++ src/StructuredParser.php
@@ -2,15 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Pgsql\Composite;
+namespace Yiisoft\Db\Pgsql;
+
+use function in_array;
 
 /**
- * Composite type representation to PHP array parser for PostgreSQL Server.
+ * Structured type representation to PHP array parser for PostgreSQL Server.
  */
-final class CompositeParser
+final class StructuredParser
 {
     /**
-     * Converts composite type value from PostgreSQL to PHP array
+     * Converts structured (composite) type value from PostgreSQL to PHP array
      *
      * @param string $value String to convert.
      */